# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/asset/finding": {
            "get": {
                "description": "Returns a paginated list of assets finding",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get All Asset Finding",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetAllAssetFinding"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create an asset finding",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Post Create Asset Finding",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Finding Category",
                        "name": "finding_category",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Finding Notes",
                        "name": "finding_notes",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Finding Image (JPG,PNG,JPEG)",
                        "name": "finding_image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Placement Id",
                        "name": "asset_placement_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseCreateAssetFinding"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/asset/finding/hour-total": {
            "get": {
                "description": "Returns a paginated list of assets finding total per hour",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get All Asset Finding Hour Total",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetFindingHourTotal"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/asset/finding/{id}": {
            "delete": {
                "description": "Permanentally delete asset finding by id",
                "tags": [
                    "Asset"
                ],
                "summary": "Delete Asset Finding By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of asset finding",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseDeleteAssetFindingById"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/asset/most-context/{targetCol}": {
            "get": {
                "description": "Returns a list of most appear item in asset maintenance by given field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get Most Context Asset Maintenance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target Column to Analyze (such as: maintenance_day)",
                        "name": "targetCol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetMostContext"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.AssetFinding": {
            "type": "object",
            "properties": {
                "asset_placement_id": {
                    "description": "FK - Asset Placement",
                    "type": "string"
                },
                "asset_placements": {
                    "$ref": "#/definitions/entity.AssetPlacement"
                },
                "created_at": {
                    "type": "string"
                },
                "finding_by_technician": {
                    "description": "FK - Technician",
                    "type": "string"
                },
                "finding_by_user": {
                    "description": "FK - User / Guest",
                    "type": "string"
                },
                "finding_category": {
                    "type": "string"
                },
                "finding_image": {
                    "type": "string"
                },
                "finding_notes": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "technicians": {
                    "$ref": "#/definitions/entity.Technician"
                },
                "users": {
                    "$ref": "#/definitions/entity.User"
                }
            }
        },
...
        "entity.AssetPlacement": {
            "type": "object",
            "properties": {
                "asset_desc": {
                    "type": "string"
                },
                "asset_id": {
                    "description": "FK - Asset",
                    "type": "string"
                },
                "asset_owner": {
                    "description": "FK - Technician",
                    "type": "string"
                },
                "asset_qty": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "description": "FK - Admin",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "room_id": {
                    "description": "FK - Room",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "entity.Metadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "entity.ResponseCreateAssetFinding": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset finding created successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseDeleteAssetFindingById": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset finding deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetAllAssetFinding": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.AssetFinding"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "asset finding fetched"
                },
                "metadata": {
                    "$ref": "#/definitions/entity.Metadata"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetFindingHourTotal": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StatsContextTotal"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "asset finding fetched"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "entity.ResponseGetMostContext": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StatsContextTotal"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "asset fetched"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.StatsContextTotal": {
            "type": "object",
            "properties": {
                "context": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "entity.Technician": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "description": "FK - User",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "telegram_is_valid": {
                    "type": "boolean"
                },
                "telegram_user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "telegram_is_valid": {
                    "type": "boolean"
                },
                "telegram_user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}